refactor: [ENGINE-2] #comment Cloud Issue Delete 부분 결과 값 리턴하도록 수정 - 23.08.08
#close #time 1h +review SR @313devops

[ Backend-Core::Java-Service-Tree-Framework(JSTF)::Advanc2d ]
[Requirement Manage] http://www.a-rms.net
[Document] http://www.313.co.kr/confluence
[IssueTracker] http://www.313.co.kr/jira
[VersionControl] https://github.com/313devgrp
[CodeReview] http://www.313.co.kr/fecru
[CICD-Deploy] http://www.313.co.kr/jenkins
[BuildManager] http://www.313.co.kr/spinnaker
[ArtifactManager] http://www.313.co.kr/nexus
[CodeAnalysis] http://www.313.co.kr/sonar
[Docker Hub] https://hub.docker.com/u/313devgrp

[Kubernetes] http://www.313.co.kr:31380
[DockerSwarm] http://www.313.co.kr/portainer/#!/auth

[DB Admin] http://www.313.co.kr/phpmyadmin/
[S3 Admin] http://www.313.co.kr/minio/login
[MEM Admin] http://www.313.co.kr/redis/
[NoSql Index] http://www.313.co.kr/elasticsearch/_cat/indices
[NoSql Node] http://www.313.co.kr/elasticsearch/_nodes?pretty=true
[Kibana] http://www.313.co.kr/kibana/app/home#/
[LogStash] http://www.313.co.kr/logstash/_node/stats?pretty
[ZipKin] http://www.313.co.kr/zipkin/
[ElasticHQ] http://www.313.co.kr/elastichq/#!/
[Grafana] http://www.313.co.kr/grafana

[NAS] http://www.313.co.kr:5000/webman/index.cgi
[Mail] http://www.313.co.kr/mail/
[Auth] http://www.313.co.kr/auth/
[RDP] http://www.313.co.kr/guacamole/#/

+++ src/main/java/com/arms/jira/cloud/jiraissue/service/CloudJiraIssueImpl.java
@@ -139,36 +139,57 @@ public class CloudJiraIssueImpl implements CloudJiraIssue {
 
     @Transactional
     @Override
-    public void deleteIssue(String connectId, String issueKeyOrId) throws Exception {
+    public Map<String,Object> deleteIssue(String connectId, String issueKeyOrId) throws Exception {
+
+    	Map<String, Object> result = new HashMap<String, Object>();
 
         String endpoint ="";
         JiraInfoDTO found = jiraInfo.loadConnectInfo(connectId);
         WebClient webClient = CloudJiraUtils.createJiraWebClient(found.getUri(), found.getUserId(), found.getPasswordOrToken());
 
-        if(checkSubTask(connectId, issueKeyOrId)){ //서브테스크가 있을 경유
-            for(int i=0;i<getSubTask(connectId, issueKeyOrId).size();i++){
+        if (checkSubTask(connectId, issueKeyOrId)){ //서브테스크가 있을 경유
+            for (int i=0;i<getSubTask(connectId, issueKeyOrId).size();i++) {
                 convertSubtaskToIssue(connectId, String.valueOf(getSubTask(connectId, issueKeyOrId).get(i).getId()), issueKeyOrId);
             }
             endpoint= "/rest/api/3/issue/" + issueKeyOrId +"?deleteSubtasks=true";
-        }else{
+        } 
+        else{
             endpoint = "/rest/api/3/issue/" + issueKeyOrId +"?deleteSubtasks=false";
         }
 
-        CloudJiraUtils.delete(webClient, endpoint, Void.class).block();
+        Optional<Boolean> response = CloudJiraUtils.executeDelete(webClient, endpoint);
 
-        cloudJiraIssueJpaRepository.deleteById(issueKeyOrId);
+        boolean isSuccess = false;
 
+        if (response.isPresent()) {
+            if (response.get()) {
+                // PUT 호출이 HTTP 204로 성공했습니다.
+                isSuccess = true;
+                result.put("success", isSuccess);
+                result.put("message", "이슈 삭제 성공");
+
+                cloudJiraIssueJpaRepository.deleteById(issueKeyOrId);
+
+                return result;
+            }
+        }
+
+        result.put("success", isSuccess);
+        result.put("message", "이슈 삭제 실패");
+
+        return result;
     }
 
     public List<CloudJiraIssueDTO> getSubTask(String connectId, String issueId){
         List<CloudJiraIssueDTO> SubTaskList = getIssue(connectId, issueId).getFields().getSubtasks();
+
         return SubTaskList;
     }
 
     public boolean checkSubTask(String connectId,String issueKeyOrId){
-        if(getSubTask(connectId, issueKeyOrId).size()>0){
+        if (getSubTask(connectId, issueKeyOrId).size()>0) {
             return true;
-        }else return false;
+        } else return false;
     }
 
     public void convertSubtaskToIssue(String connectId, String  subTaskKeyOrId,String issueKeyOrId) throws Exception {
@@ -201,7 +222,7 @@ public class CloudJiraIssueImpl implements CloudJiraIssue {
 
     @Transactional
     @Override
-    public String collectLinkAndSubtask(String connectId) {
+    public Map<String,Object> collectLinkAndSubtask(String connectId) {
         List<CloudJiraIssueEntity> list = cloudJiraIssueJpaRepository.findAll();
         List<CloudJiraIssueEntity> saveList = new ArrayList<>();
 
@@ -251,6 +272,10 @@ public class CloudJiraIssueImpl implements CloudJiraIssue {
             cloudJiraIssueJpaRepository.saveAll(saveList);
         }
 
-        return "success";
+        Map<String, Object> result = new HashMap<String, Object>();
+        result.put("success", true);
+        result.put("message", "스케줄러 작동 완료되었습니다.");
+
+        return result;
     }
 }
